+++ test/typescript/apis/project/cds-services.ts
@@ -129,6 +129,13 @@ await srv.send('READ', 'Books')
 await srv.send('boundAction', 'Books', { book: 251, quantity: 1 })
 await srv.send('unboundAction', { book: 251, quantity: 1 })
 
+// schedule + flush
+await srv.schedule({ method: 'READ', path: 'Authors' }).after(1000 /* ms */)
+await srv.schedule({ query, headers: {} }).every(2, 's')
+await srv.schedule({ event: 'READ' }).after('1h')
+await srv.schedule('CREATE', 'Books', {}, {}).every('3d')
+await srv.flush()
+
 // TX
 let tx = cds.tx({})
 tx = cds.transaction({})